benchmarks/fapws3/requirements.txt
gunicorn==19.1.1
meinheld==0.5.6
uwsgi==2.0.7
Werkzeug==0.9.6
waitress==0.8.9

six==1.8.0

fapws3==0.11.dev

newrelic